requirement.txt

Django==4.1.3
keras==2.11.0
matplotlib==3.6.2
mediapipe==0.9.0
numpy==1.23.5
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
playsound==1.2.2
python-dateutil==2.8.2
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.28.0
threaded==4.1.0
h5py==3.7.0
Pillow==9.3.0